+++ client/__stories__/components/common/Header.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from '@/components/Header';
+import Header from '@/components/common/Header';
 import { ComponentStory } from '@storybook/react';
 
 export default {
