+++ client/__stories__/components/common/Input.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Input from '@/components/Input';
+import Input from '@/components/common/Input';
 import { ComponentStory } from '@storybook/react';
 
 export default {
