+++ client/__stories__/components/common/Modal.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Modal from '@/components/Modal';
+import Modal from '@/components/common/Modal';
 import { ComponentStory } from '@storybook/react';
 
 export default {
