+++ client/__stories__/components/signup/ImgSelector.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ImgSelector from '@/components/ImgSelector';
+import ImgSelector from '@/components/signup/ImgSelector';
 import { ComponentStory } from '@storybook/react';
 
 export default {
