+++ client/__tests__/components/common/SearchInput.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SearchInput, { SearchInputProps } from '@/components/SearchInput';
+import SearchInput, { SearchInputProps } from '@/components/common/SearchInput';
 import { fireEvent, render, RenderResult } from '@testing-library/react';
 
 const DEFAULT_ARGS: SearchInputProps = {
