+++ client/__tests__/components/signup/ImgSelector.test.tsx
@@ -1,7 +1,7 @@
 import { MockImage } from '@mocks/MockComponent';
 
 import React from 'react';
-import ImgSelector, { ImgSelectorProps } from '@/components/ImgSelector';
+import ImgSelector, { ImgSelectorProps } from '@/components/signup/ImgSelector';
 import { fireEvent, render, RenderResult } from '@testing-library/react';
 
 jest.mock('next/image', () => MockImage);
