+++ client/src/components/common/SearchInput.tsx
