+++ client/src/pages/index.tsx
@@ -1,8 +1,5 @@
 import React from 'react';
-import type { NextPage } from 'next';
 
-const Home: NextPage = () => {
-  return <div>MainPage</div>;
-};
-
-export default Home;
+export default function MainPage(): JSX.Element {
+  return <>Main Page</>;
+}
